+++ examples/prometheus_cache/check_many_metrics.py
@@ -17,8 +17,9 @@ import time
 from typing import List
 
 from sarc.client.job import get_job
+
 from sarc.config import scraping_mode_required
-from sarc.jobs.series import get_job_time_series
+from sarc.core.scraping.series import get_job_time_series
 
 logger = logging.getLogger(__name__)
