+++ sarc/core/scraping/prometheus.py
@@ -6,11 +6,10 @@ from sqlmodel import Session, col, select
 
 from sarc.cache import Cache, CacheEntry
 from sarc.config import ClusterConfig, config
+from sarc.core.scraping import series
 from sarc.db.cluster import SlurmClusterDB
 from sarc.db.job import JobStatisticDB, SlurmJobDB
 from sarc.db.runstate import get_parsed_date, set_parsed_date
-from sarc.jobs import series
-from sarc.jobs.series import JOB_STATISTICS_METRIC_NAMES
 from sarc.models.job import SlurmState
 from sarc.traces import trace_decorator
 
@@ -50,7 +49,7 @@ def fetch_prometheus(
     with cache.create_entry(datetime.now(UTC)) as ce:
         for entry in sess.exec(query):
             raw_prom_data = series.get_job_time_series_data(
-                job=entry, metric=JOB_STATISTICS_METRIC_NAMES, max_points=10_000
+                job=entry, metric=series.JOB_STATISTICS_METRIC_NAMES, max_points=10_000
             )
             if raw_prom_data == []:
                 continue
