--- sarc/jobs/__init__.py
@@ -1,3 +1,0 @@
-from .series import get_job_time_series_metric_names
-
-__all__ = ["get_job_time_series_metric_names"]
