+++ scripts/prometheus/backup.py
@@ -13,7 +13,7 @@ from tqdm import tqdm
 from sarc.config import UTC, config
 from sarc.core.models.validators import datetime_utc
 from sarc.core.scraping.jobs_utils import update_allocated_gpu_type_from_nodes
-from sarc.jobs.series import (
+from sarc.core.scraping.series import (
     JOB_STATISTICS_METRIC_NAMES,
     _get_job_time_series_data_cache_key,
     _get_job_time_series_data_cache_subdir,
